Replace use of unstructured
This is a simplification that's been on my TODO list for a while now.
Someone else was working in this code today (you know who you are) and
it seems like a good time to start improving this.

This code has accumulated many layers over the last year, and some of
them are very much unnecessary.

---

Primarily the issue fixed here is that we were using a Kubernetes type
(`unstructured.Unstructed`) in the core of the system rather than our
own type (`GenericComponent`). This is a holdover from a very early
stage of the project, and hopefully this change makes things just that
much clearer and easy to extend.

My plan for this area is that we should do a lot of this kind of work
via templating or code generation eventually. For now I'm introducing
some stronger types.