Update our SDK references to latest
This came up in PR feedback - it's better for us to float along on
recent API versions than to pin to a specific version (which might be
old). In general the API import paths for the azure SDK should not
contain dates for us (unless we have a good reason).