Add some basic schemas
I spend some time during meetings putting together schemas for a few
types we accept today. The idea is that we'll use these schemas for:

- Bicep intellisense
- SDK generation
- Validation (of payloads)
- Generating types for our code

We'll need to really *scale up* our ability to work with new data types,
and probably invest in a lot of our own logic and tooling around these
schemas.

I wanted to put something basic out there that establishes the pattern
so we can maintain these and add more as we light up new data types.

The work to start using these is scheduled to start relatively soon, and
I think we're at the tipping point where we should think about becoming
more data driven.