Use type `Proxy` instead of `Proxy,Cache`
This changes how we're registering our resource types with ARM. What
this means that GET requests (GET/LIST) will be sent through to our RP
rather than being handled by ARM's cache on our behalf.

The reason for this change is that `Proxy,Cache` is resulting in our
top-level `kind` property being omitted. We return this property from
our RP, but it's not in the final response :-/.

We're following up with the ARM team on this behavior.