Add "channel" concept to our asset publishing
Part of: #231

This change makes our version publishing logic more sophisticated. After
this change we produce two version from each tagged release:

- The release version (full version)
- The channel (excludes patch)

Things that get a version number baked in are stamped with the release
version. That way a `rad --version` output will include the full build
info including the patch level.

Things that are downloaded by the user are uploaded based on a
"channel". This allows us to patch things and for users to see the
update. For instance we can patch the RP by uploading a new build with
the same tag. Users can restart their RP and pick up the update.

There's also built-in support for updating a version marker to point to
the latest channel. This way we can push a `v0.2` tag and then `0.2`
becomes the stable channel. Our install scripts and other assets will
refer to this channel through the file.