Add support for ingress & inbound route trait
Part of: #15

This change adds support for the InboundRoute trait, which maps to a
Kubernetes ingresss (ingress-nginx).

Note that this change requires Kubernetes 1.19. Using these features in
your environment will require that as the minimum version as well as an
install of ingress-nginx (now part of environment setup).

This change does not include an integration test. That will come in a
follow up as the test environments need to be patched first.