Fix retries of failed deployments
Fixes: #375

Currently we treat a deployment with no changes as a *noop* and
immediately response with a 200. This is a problem when the previous
attempt to deploy ended in a non-successful state. We respond with a
200+non-success state which ARM considers a protocol violation.

This change adds logic to force a redeploy when the previous deployment
ended in a failure state. Adding/removing components in an update to the
deployment will still work as before but for any components that are
unchanged we'll *force* and update so they are retried.