Tolerate 404 on deletion of Cosmos database
Fixes: #22

This change applies the pattern we're using for service-bus to deletes
of cosmos db. Notably we allow deletion to succeed when we get a 404 on
the database (a nested resource).

The problem this solves is making sure a retry of a delete can succeed.
The current state is that a delete of a deployment might fail partway
though. If the CosmosDB Account is deleted before deletion fails, then
we are currently unable to delete the deployment on retry. This handles
the case where the account is already gone.

This is a special case for nested resources. Top level resources return
a 204 for delete regardless of whether they exist. However, *certain*
nested resources will return a 404 if the containing resource does not
exist. This is the last remaining case I'm aware of where we don't
handle this properly.

In the future when we generalize working with Azure resources we will
generalize this as well.