Add support for private container registry at init
This change adds support for initializing an environment with an ACR
registry as part of `rad env init azure`. This is a stopgap for 0.2 to
make ACR registries easy to use. We'll do a more thorough design pass on
this in the future and produce a more flexible and declarative approach.