Fix issues found during tutorial walkthroughs
Fixes: #473
Fixes: #472

This change addresses small issues that I found while walking through
the tutorial, as well as updates the out of date packages used by the
webapp tutorial.

I found a bug where step 2 (add db) of the webapp tutorial did not
function properly, but I think the issue is that the image was out of
date. I added some logging to this for debugging, but the problem when
away when I updated the image.