Refactor appmodel into a separate package (#494)
* Refactor appmodel into a separate package

This is part of a series of changes to bring Kubernetes support into the
main codebase of Radius and out of my prototype branch.

This change refactors the application model into a separate package that
is decoupled from the deployment processor. This includes the set of
known component types and resource handlers.

The value of this change is that now the deployment processing is
decoupled and data-driven. It is told about the set of supported
resources and types, and then acts accordingly. Now it's possible to
create different 'models' with different sets of supported component
kinds. In the future this will also include details like traits, which
will allow us to vary support across different hosting platforms.

* Decoupled deployment processor for Kubernetes

* Rename folders for uniqueness in docs

* Patch namespace in all kubernetes paths

* one more fix