Separate control-plane and application resource groups
Fixes: #303

This change addresses the original scope of #303 - that the control
plane should be a separate resource group from the radius resources and
application resources that the user interacts with. This is a closer
approximation of what a real service based on Radius would look like to
a user (if you close your eyes and ignore the control plan group).

The change touches 4 major areas:

- The init process - we now need to create two resource groups
- Environments and CLI - management and deployment interact with the
  'app' resource group and diagnostics interacts with a k8s in the
  'control plane' resource group
- The RP - we need to interact with one group for deploying Azure
  resources and a second group for interacting with k8s.
- Docs - we need to explain ourselves...

Writing the ARM template for this was a real slog, and I had to learn
a ton about different edge cases in ARM. The main thing to understand
based on this template is that we now do a 'subscription' scoped
deployment, which allows us to create resource groups as part of the
template. Then we fire off two nested deployments, one for each group.
I've tried to improve the comments and docs as part of this change, but
it will be very hard to review. I also removed some parameters that were
totally unused.

The other changes are actually really straightforward and are hopefully
intelligible based on just the deltas in code.