Implement formatting support for CLI
Fixes: #165

This change decouples the CLI from the output produced by commands like
`rad application list`. The objects returned from the client now flow
through a configurable formatting pipeline.

We have a new global command line option `-o` for choosing output.

The main addition here is support for tabular formats with `-o table`.
We have support for a per-type set of columns to display, and can add
extensible column support in the future.

The default has changed to use table output by default. Tables are for
humans, JSON is for computers. The raw JSON output is accessible with
`-o json`.