Stop using viper.GetViper()
Found this issue while doing some test cleanup. Right now we're using
viper.GetViper() which relies on global mutable state and is thus not
testable.

I was trying to figure out why our test code that reads config doesn't
do it the way we do in `rad`... and yeah that's why....