Add consistent tags to Azure resources
Fixes: #26

I needed to do this to make the Azure resource work more testable. For
instance, it's not straightforward for us to run queries on Azure
resources because they don't have stable names. Tags help a lot with
this as well as for general troubleshooting.

While I was here I also cleaned up and made consistent some of the
handling of Kubernetes labels. We're labeling things related to the
container in a pretty detailed way and we want to do the same with other
resources. Again, this is super useful for testing and just general
troubleshooting.

I consolidated all of this stuff in a higher level package with just the
label/tag functionality. Since this is really generic stuff and has no
dependencies on other packages it should be a leaf node. For example I
couldn't leave the label constants where they are today because it would
create a cycle.