Prevent crosstalk between test runs
While looking at PR #616, I noticed that we haven't kept our test run
timeout and our environment lease timeout in sync. Our environment lease
timeout is 30 mins and our test run timeout is 60 minutes.

That means that any time a test run has an environment in use for 30
minutes, there's a risk that another test run will start using it as
well. This can of course lead to double-deletion.

The fix for these is to keep these two values in the same file
`build.yaml` and make sure sure that the lease expiry time is pretty
generous.