Remove the "Error:" prefix from error message and use encoding/json
Usually when printing error messages, the caller of
GenerateErrorMessage (or owner of the err) will have more context
so it's better to just leave them to prepend the prefix later.

The tests may seem redundant but they are useful to confirm that the
deserialized information is enough to be parsed, and thus not containing
silly pointers printouts.