Add management client for Radius Kubernetes (#773)
Fixes: #760

This change adds support for interacting with Kubernetes via `rad
component list` and similar commands. For now this support is based on a
conversion to the types we return from ARM. This will result in the same
experience we have for Azure environments.

In the future we might want to introduce a *format neutral*
representation for use this CLI. I didn't puruse that in this PR because
we're currently having active discussions about the shape of the model
that will impact this.