Add support for placeholders (#850)
This change adds support for placeholders as part of a resource
dependency, this will allow you to copy values from a dependendent
output resource into the body of another resource.

I'm adding the basic plumbing in this PR and kachawla will wire it up to
the deployment processor.

As a recent example of this take a look at how Kubernetes support for
MongoDB works. It 'renders' a Service, Secret, and StatefulSet. The
names of the Service and Secret need to get baked into the `spec` of
the  StatefulSet. This placeholder functionality is for expressing this
kind of dependency.