Standardize pattern for LRO deletion (#881)
Fixes: #880

This change makes all of our *long-running* resource deletion code
consistent and fixes some bugs.

This is the result of finding a crash in tests when these Azure SDK
calls return an error. In general the issue here is that the existing
code inspects the *non-error* return value when an error is returned.
There's no guarantees about the state of the object when that happens.
We also need to make sure to *early-exit* the function when a 404
happens, the code was continuing, which was a bug.

Also moved this to `pkg/azure/` per-discussion for code reorganization.