Refactor: decouple services inside the RP (#998)
This is a refactor that decouples the multiple services that run inside
the resource provider. Right now we have three different things that run
in the same process and interact over channels. This change makes the
startup path more straightforward/standardized and decouples how these
services start.

Also added here is a lot better resiliancy and a better pattern for
adding more of these services to the current monolith. The mechanics of
starting/stopping are contained now to a 'host' abstraction.