Add v3 Radius RP abstraction and plumbing (#1015)
* Remove dead code

* Separate handler from server

* Restrict v2 rp to 'radius' custom rp name

* Add v3 routes

* Rough in v3 RP and hook up to REST API

This change adds the v3 RP interface and wires it up to the REST API via
the router and the `radiusv3` Custom RP name. This ensures that there
will be no crosstalk between the two different versions of the RP.

This change just contains some of the necessary plumbing, there's no new
business logic here.