Add adapter to renderers
This change adds some missing things for the computed value construct
that we discussed previous.

Also this change demonstrates how we can adapt a subset of the  existing
v1 renderers to use the new renderer contract and code path without
rewriting it as a transitional measure. I can adapt the rest of these in
a follow-up change after we review the pattern.

This is preferrable to doing a 'big bang' rewrite because because we'll
be able to migrate each scenario individually.