Add AppModelv3 Container Renderer (#1059)
This change adds a copy of the container renderer that's updated to
AppModel v3's semantics. This builds on the work that jkotalik has done
in parallel but also merges in functionality that was merged to main
since that work started.

I used the v1 tests as the starting point to keep us honest here.