Fix error handling and context in the RP v3
This change fixes two issues:

We weren't handling the case where an operation is in a terminal fail
state without a specific error. Now we just return a generic error
message when that happens.

We also need to create a new context to pass to the deployment
processor. The context we're using when handling an HTTP request is tied
to the request's lifetime, and will be canceled when the request
completes. So we need a background context to hand off for background
processing.