Report errors inside deployment processor
This change improves error reporting for failures that occur during
deployment/deletion. We need to propagate these errors to the operation
so that they can be reported to an async client.

The current code will report a 500 with a generic error message for all
kinds of failures, this plumbs through the actual detailed error
messages.