Enablement of the first AppModelv3 scenario (#1062)
* Support connections in magpie

This updates the magpie test application to support the environment
variables we create by default in AppModelv3

* Log HTTP status codes when responding

We're not making quite enough noise in the logs

* Fix incorrect logging of shutdown

We were being silent about errors that occur during startup and logging
it as a graceful shutdown instead. I hit this while testing locally, and
ran into a port conflict which wasn't reported clearly.

* Log unhandled errors reported by the RP

Nothing was logging these errors, this code path catches all of them

* Fix bugs in the database implementation

These APIs require a reference/pointer and were being passed a struct
directly.

* Add container and servicebus to v3 model

These renderers have already been updated to work with v3, adding them
so they can be used.

* Update LocalRP client to work with v3

This change makes our 'localrp' test client work in a more generic way
instead of only handling a hardcoded set of resource types.

This also handles property references between resources which is
necessary for AppModelv3 due to resource ids as references.

* Fix error handling and context in the RP v3

This change fixes two issues:

We weren't handling the case where an operation is in a terminal fail
state without a specific error. Now we just return a generic error
message when that happens.

We also need to create a new context to pass to the deployment
processor. The context we're using when handling an HTTP request is tied
to the request's lifetime, and will be canceled when the request
completes. So we need a background context to hand off for background
processing.

* Report errors inside deployment processor

This change improves error reporting for failures that occur during
deployment/deletion. We need to propagate these errors to the operation
so that they can be reported to an async client.

The current code will report a 500 with a generic error message for all
kinds of failures, this plumbs through the actual detailed error
messages.

* Fix output of application rest responses

We were outputting the location as the property "omitempty" instead of
as the location.