Update servicebus test to use AppModel v3 (#1064)
This changes gets our first e2e test working (managed servicebus). I'm
still waiting on the CLI changes so there's no deletion for right now.

I introduced some simplications to the health code to make it work
across v1 and v3. We're passing in the resource ID of the 'owner'
resource for health monitoring rather than a bunch of granular
information. This makes the health system less coupled to the structure
of our resources which is a nice improvement.