Add HttpRoute and manual scale support
This change ports support for jkotalik's work on Kubernetes and makes
some improvements to how services and pods get matched up.

The bigger changes here:

- Port manual scaling to v3
- Simplify how gateways/ingress work for v3

Gateways (public access) are now part of the HttpRoute rather than a
trait on the ContainerComponent. This is much simpler to write and also
simpler to code up.

Additionally updates the main body of 'container' tests to use
AppModelv3.