Implement exchange type for resource identity
Ok soooooo....

This change started because I found an issue with our handling of
'resource identity' in the database. We were persisting the
'outputResourceInfo' as `interface{}` and the result is that it doesn't
round-trip. So while we were generally *writing* 'outputResourceInfo',
nothing reads it. I need to read it for the next thing I'm building.

Chasing this problem led me to the contract with the health system, and
the realization that internal details of the health system ('healthId')
had leaked out all the way into our REST API. Right now we're passing
about 5 identifiers from the RP to the health system, and ultimately
that should be 2:

- The identity of the underlying resource
- The identity of the 'owning' Radius resource

The rest of this change is cleanup that comes from pushing that
realization upstream into the health code. One of the other things that
I entangled along the way, is the coupling of two contracts:

- The contract for communicating between RP and Health
- The contract for communicating *inside health*

There was a lot more refactoring done than I initially set out to do but
I think this positions us well to move the health system into Kubernetes
provided that we all deem these to be positive changes.