Support listSecrets custom action (#1121)
This change adds support for an ARM 'custom action' that will pass
through the CustomRP infrastructure into our RP.

I've added support to the compiler to allow *programmatic access to
secrets*. ex: `foo: db.connectionString()`. Where our component types
support secret values, the Bicep compiler will inject a function symbol
that maps to this new custom action. This is similar to how handling of
secrets generally works in ARM, but more streamlined.

The next change after this one will add the RP functionality to
implement this. It's good to separate manifest changes from the code
changes, because the manifest is not applied by the CI.