Add support for programmatic access to secrets
This change adds support for programmatic access to secrets from ARM
using a function call. I've already added support for this in the Bicep
compiler, and it will translate the function call into the appropriate
call into the RP. This change adds the backend support and a test.

Example (db is a `mongodb.com.MongoComponent`):

`FOO: db.connectionString()`

This will resolve to a POST against our RP, which the changes in this PR
handle.