Add secrets handling and update MongoDB/CosmosDB (#1108)
This change introduces support for late-binding to secrets. The renders
capture/describe how to access the secrets and the control plane can
look them up later in a loosely coupled way. This is generic across any
ARM resource.

I have not yet implemented Kuberentes support for secrets but the rough
sketch is there.

This PR also updates both flavors of CosmosDB and MongoDB to use
appmodelv3 and removes the older versions. The MongoDB version had to be
migrated at the same time, but will be blocked on secret support for
Kubernetes.

I was able to reuse a lot of the improvements that were recently made to
the CosmosDB mongo code path and port them to CosmosDB SQL. Now they
share the same structure and a lot of code.