Add support for programmatic secrets to K8s
Fixes: #595

This change adds support for programmatic secret access to the K8s
deployment engine. This means an expression like `db.connectionString()`
will result in a secret-based lookup for one of the radius types. We
don't support any other kind of 'custom actions' at this time other than
parity with what we do for the Radius RP.