Add support for secret values in connection in k8s (#1132)
This change enables the use of secrets in Kuberentes from inside the
renderer. The current code only provides the renderer access to the
non-secret values, so secrets would be omitted.

Also fixed an issue with how we were transforming the 'defintion' when
passing into the the renderer and added a unit test.