Update Managed Identity+Role Assignment for V3 (#1145)
This change enables the IAM functionality for AppModelv3. In addition to
enabling this for KeyVault I also made updates to generalize these code
paths so they can be applied to more resource types in the next
milestone.

The existing code had a lot of assumptions in place about the set of
constrained scenarios where it would need to be used:

- Only with keyvault
- Fixed set of roles being assigned

This builds on the work that Karishma did previously in this area to
decouple the creation of these resources.