Scrub improper usage of 'component' and dead code (#1218)
This is a pass over all usage of the term 'component' in our go code to
ensure that we're not misusing it. Component is now a category of things
we manage, so it's sometimes the right term to use and sometimes not.

I also found some more dead code as part of the process (unused
renderers that are obsolete).