Use sub and resource group for unique names (#1600)
Fixes: #1596

This change adds the subscription id as a seed value for generating
unique names in our environment deployment template. Previously we were
just using the environment name (same as the resource group) but not
considering the subscription. The result is that environments with the
same name across different subs would collide.