Implement progress (spinners) for deployment (#1572)
This change implements a progress notifications for deployments as part
of the deployment client as well as a console UI that shows while
operations are in progress and which are completed.

The tricky thing here is that we need to show multiple 'bars' or
'spinners' since we have concurrent operations. No existing go library
with an OSS licence handles this well. The libraries out there either
support a single 'spinner' or have other limitations that would be
unacceptable. For this reason I had to go somewhat lower level and use a
dependency of one of the 'spinner' libraries to do what we want.