Add preliminary 'builder' support (#1594)
Fixes: #1556

This change adds preliminary support for builds as part of 'rad.yaml'.
Users can describe build tasks as part of 'rad.yaml' and those build
tasks will run as part of deployment. In particular the build results
can be consumed by IAC code as a parameter.

This change is preliminary and adds the general framework along with a
**very** simple builder integration for docker/Dockerfiles. The
intention here is that we will grow a plugin ecosystem of 'builders'
that can be maintained by the community and is loosely coupled to
Radius. We need build support for inner-loop experiences that work with
`rad`.