Perform build in parallel for rad deploy
This change adds build parallelism (and colorized output) to rad app
deploy and friends. This is a significant improvement in performance
that scales nicely, especially given than a "no op" container build
and push takes at least 5-10 seconds. For the container apps sample I've
been using this is a ~20s improvement for me when all of the images are
up to date already.

The bulk of the work here is in adjusting the output to be color-coded
and prefixed. I was able to find some good examples to build on here.
The principle is that we implement infra for color-coded and named
output streams, and then layer an `io.Writer` implementation on top.