Use digests for versioning in docker build
This change adds digest support for `rad deploy` and friends. By using a
digest for versioning in means that you can iterate safely regardless of the
tagging strategy in use, since the digest always refers to the exact
image that was produced.

This has a similar effect to using `imagePullPolicy` except its way more
deterministic. `imagePullPolicy` requires us to programmatically restart
all of the running pods. With digest versioning the manifest actually
changes build-to-build and so we don't need to treat it as a special
case; it's just another deployment.