Use local registry for push/pull in local dev
This change adds support for utilizing the local container registry
created by k3d. We already provision the registry during environment
setup for local dev, and this change wires it up.

To accomplish this we add a generic 'registry override' capability to
environments. This allows a user to specify the registry they want to
use for producting in code and treat the local registry as an
implementation detail of radius. The goal is that users don't even need
to know this is happening, things just go fast.

We will likely need to expand this functionality in general for
registries/environments in the future for Radius as a whole - for
example better support for private registries. Just making a note for
posterity that this PR adds a specific feature related to registries,
but there will be more to do here.

The main detail in this change to be aware of is that registry push/pull
is asymmetric when it comes to local dev. You don't use the same URL to
push as you do pull, since the networking behaviors (DNS) are different
in each case.