Add hardening to kubernetes watch code
We're seeing a lot of irrelevant spam in logs, and also some cases where
tests are failing due to 'not a pod' messages.

My goal with this PR is to clean up the logs such that we're not logging
things unless they are actionable. Currently we log things like
'unexpected EOF' or 'context cancelled' that should be suppressed.